Fixed a crash while reading the match file
If the machine's code page is not 65001(utf-8) and there's a character other than ASCII inside the image_path, the std::filesystem::path::operator=(const char*) crashes if the given string is utf-8.
It would have been the case for the config file, too, since it also uses paths.
Forced it to use the utf-8 encoding by reinterpret_casting to const char8_t. Not confirmed if it'll work if it's compiled with libstd since this behavior is not specified in the standard.