Sending VK_LSHIFT and VK_RSHIFT in addition to VK_SHIFT
Fixing bug #1.

Apparently VK_SHIFT is treated as the same as VK_LSHIFT, and VK_RSHIFT is considered differently even though the docs says VK_SHIFT is a key for all.

Found another potential bug while fixing this.
If the letter doesn't need a shift key, should check if it it pressed or not and if so, need to send KEYUP event before sending the letters and KEYDOWN event at the end.
Will close the issue after fixing this too.