Correctly handling left shift and right shift
Final fix on bug #1.

Unpressing the shift key if the letter to be sent doesn't need a shift and pressing again after sending.

So, it seems sending all shift keys was a bad idea since the left key is stuck if you were using the right key. Ugh. Handling the two separately and never using VK_SHIFT when it comes to unpressing and pressing again.